+++ src/vga.rs
@@ -1,12 +1,14 @@
 use crate::{
+    interrupts,
     ports::{self, Port},
+    sysinfo::SystemInfo,
     wrappers::UnsafeFlag,
 };
 use core::{
     convert::Infallible,
     fmt::{self, Write},
     ptr,
-    sync::atomic::{AtomicU8, Ordering},
+    sync::atomic::{AtomicBool, AtomicU8, Ordering},
 };
 
 /// The color palette used by `VGAChar`
@@ -41,6 +43,24 @@ impl VGAChar {
     /// The space character.
     const SPACE: VGAChar = VGAChar::new(0x20, Color::White, Color::Black);
 
+    /// A box character for top left corners.
+    pub const TOPLEFT_CORNER: u8 = 0xC9;
+
+    /// A box character for bottom left corners.
+    pub const BOTTOMLEFT_CORNER: u8 = 0xC8;
+
+    /// A box character for top right corners.
+    pub const TOPRIGHT_CORNER: u8 = 0xBB;
+
+    /// A box character for bottom right corners.
+    pub const BOTTOMRIGHT_CORNER: VGAChar = VGAChar::new(0xBC, Color::Grey, Color::Black);
+
+    /// A box character for vertical borders.
+    pub const VERTICAL_BORDER: VGAChar = VGAChar::new(0xBA, Color::White, Color::Black);
+
+    /// A box character for horizontal borders.
+    pub const HORIZONTAL_BORDER: VGAChar = VGAChar::new(0xCD, Color::White, Color::Black);
+
     /// Constructs a new color using `fg` as the text color and `bg` as the background color.
     pub const fn new(char: u8, fg: Color, bg: Color) -> VGAChar {
         VGAChar((char as u16) | (bg as u16) << 12 | (fg as u16) << 8)
@@ -64,7 +84,7 @@ impl YoinkedBuffer {
     /// Tries to return a mutable reference to the buffer.
     ///
     /// Fails if the buffer is being used somewhere else.
-    fn try_yoink() -> Option<Self> {
+    pub fn try_yoink() -> Option<Self> {
         if !BUFFER_HELD.load() {
             // Safety: BUFFER_HELD is private to YoinkedBuffer, and the
             // check above ensures that there'll probably be only one copy of BUFFER
@@ -78,7 +98,7 @@ impl YoinkedBuffer {
     }
 
     /// Returns a mutable reference to the buffer.
-    fn buffer(&mut self) -> &mut RawBuffer {
+    pub fn buffer(&mut self) -> &mut RawBuffer {
         self.0
     }
 
@@ -109,6 +129,10 @@ pub static mut BUFFER: &mut RawBuffer = &mut YoinkedBuffer::empty_buffer();
 /// YoinkedBuffer will assume it has complete access to `BUFFER` when this static is cleared.
 pub static BUFFER_HELD: UnsafeFlag = UnsafeFlag::new(false);
 
+/// Allows printing to row 0 if set.
+/// Used to prevent overwriting topbar.
+static ALLOW_ROW_0: AtomicBool = AtomicBool::new(false);
+
 /// The VGA's current cursor info.
 ///
 /// Access this static via the `CursorPos` associated functions,
@@ -135,13 +159,18 @@ pub enum CursorShift {
 /// Prints to the vga text buffer.
 #[macro_export]
 macro_rules! print {
-    ($($args:tt)+) => ($crate::vga::_print(format_args!($($args)+)));
+    (fg = $fg:ident, bg = $bg:ident, $($args:tt)+) => ($crate::vga::_print(format_args!($($args)+), $crate::vga::Color::$fg, $crate::vga::Color::$bg));
+    (fg = $fg:ident, $($args:tt)+) => ($crate::print!(fg = $fg, bg = Black, "{}", format_args!($($args)+)));
+    (bg = $fg:ident, $($args:tt)+) => ($crate::print!(fg = White, bg = $fg, "{}", format_args!($($args)+)));
+    ($($args:tt)+) => ($crate::print!(fg = White, bg = Black, "{}", format_args!($($args)+)));
 }
 
 /// Prints to the vga text buffer with a trailing newline.
 #[macro_export]
 macro_rules! println {
     () => ($crate::print!("\n"));
+    (fg = $fg:ident, $($arg:tt)+) => ($crate::print!(fg = $fg, bg = Black, "{}\n", format_args!($($arg)+)));
+    (bg = $bg:expr, $($arg:tt)+) => ($crate::print!(fg = White, bg = $bg, "{}\n", format_args!($($arg)+)));
     ($($arg:tt)+) => ($crate::print!("{}\n", format_args!($($arg)+)));
 }
 
@@ -170,17 +199,16 @@ impl CursorPos {
         let (row, col) = Self::row_col();
 
         // Clamp row
-        if row >= BUFFER_HEIGHT {
-            println!("VGA cursor row contains invalid value of {row}! Clamping...");
-            CURSOR.row.store(BUFFER_HEIGHT - 1, Ordering::Relaxed);
-        }
+        let row = if ALLOW_ROW_0.load(Ordering::Relaxed) {
+            row.min(BUFFER_HEIGHT - 1)
+        } else {
+            row.clamp(1, BUFFER_HEIGHT - 1)
+        };
+        CURSOR.row.store(row, Ordering::Relaxed);
 
         // Clamp column
-        if col >= BUFFER_WIDTH {
-            println!("VGA cursor column contains invalid value of {row}! Clamping...");
-            let max = BUFFER_WIDTH - 1;
-            CURSOR.column.store(max, Ordering::Relaxed);
-        }
+        let col = col.min(BUFFER_WIDTH - 1);
+        CURSOR.column.store(col, Ordering::Relaxed);
     }
 }
 
@@ -194,26 +222,26 @@ pub enum Corner {
     BottomRight = 0xb903e,
 }
 
-/// Used by `_print` to write call `print_color`.
-struct VGAWriter;
+/// Used by `_print` to print.
+/// Uses `fg` as the text color and `bg` as the background color.
+struct VGAWriter {
+    fg: Color,
+    bg: Color,
+}
 
 impl Write for VGAWriter {
     fn write_str(&mut self, s: &str) -> fmt::Result {
-        print_color(s, Color::White);
+        for byte in s.bytes() {
+            write_char(byte, self.fg, self.bg);
+        }
         Ok(())
     }
 }
 
 /// Used by `print!` and `println!` to write to the VGA text buffer.
-pub fn _print(args: fmt::Arguments) {
-    write!(VGAWriter, "{args}").unwrap()
-}
-
-/// Prints `s` using `fg` as the text color and black as the background color.
-pub fn print_color(s: &str, fg: Color) {
-    for byte in s.bytes() {
-        write_char(byte, fg, Color::Black);
-    }
+pub fn _print(args: fmt::Arguments, fg: Color, bg: Color) {
+    let mut writer = VGAWriter { fg, bg };
+    write!(writer, "{args}").unwrap()
 }
 
 /// Writes `byte` to VGA as a character using `fg` as the text color and `bg` as the background color.
@@ -255,7 +283,7 @@ fn newline() {
                 buf[row as usize - 1] = buf[row as usize]
             }
 
-            // clear last row
+            // Clear the last row
             for col in 0..BUFFER_WIDTH {
                 buf[BUFFER_HEIGHT as usize - 1][col as usize] = VGAChar::SPACE
             }
@@ -267,9 +295,9 @@ fn newline() {
 
 /// Updates the position of the vga cursor based on the `CURSOR` static.
 pub fn update_vga_cursor() {
+    CursorPos::clamp_row_col();
     let (row, col) = CursorPos::row_col();
     let pos = row as u16 * BUFFER_WIDTH as u16 + col as u16;
-    CursorPos::clamp_row_col();
 
     // Safety: The cursor is forced into valid values in the line above
     unsafe {
@@ -322,7 +350,7 @@ pub fn shift_cursor(direction: CursorShift) {
             }
         }
         CursorShift::Up => {
-            if row == 0 {
+            if row == 1 {
                 CursorPos::set_row(BUFFER_HEIGHT - 1);
             } else {
                 CursorPos::set_row(row - 1)
@@ -341,7 +369,7 @@ pub fn shift_cursor(direction: CursorShift) {
 /// Fills the VGA text buffer with spaces and resets the cursor position.
 pub fn clear() {
     CursorPos::set_col(0);
-    CursorPos::set_row(0);
+    CursorPos::set_row(1);
     update_vga_cursor();
 
     // Clear the buffer
@@ -383,7 +411,35 @@ pub fn init() -> Result<(), Infallible> {
 
     // Print welcome message
     clear();
-    print!("Hello, ");
-    print_color("Sunflower!\n\n", Color::LightCyan);
+    print!("\nHello, ");
+    println!(fg = LightCyan, "Sunflower!\n");
     Ok(())
+}
+
+/// Draws the topbar with `title` as it's title.
+/// Title must be exactly 9 characters long.
+pub fn draw_topbar(title: &'static str) {
+    // Print at the top left corner
+    interrupts::cli();
+    let (prev_row, prev_col) = CursorPos::row_col();
+    ALLOW_ROW_0.store(true, Ordering::Relaxed);
+    CursorPos::set_row(0);
+    CursorPos::set_col(0);
+
+    // Do the printing
+    let sysinfo = SystemInfo::now();
+    print!(
+        fg = Black,
+        bg = Cyan,
+        " {} on {} | {title} | Help: SysRq / PrntScr F7 | {}",
+        sysinfo.sfk_version_short,
+        sysinfo.cpu_vendor,
+        sysinfo.patch_quote
+    );
+
+    // Restore previous vga state
+    ALLOW_ROW_0.store(false, Ordering::Relaxed);
+    CursorPos::set_row(prev_row);
+    CursorPos::set_col(prev_col);
+    interrupts::sti();
 }